0.1.3 Add CI/CD workflows, dev scripts, docs, and update deps
Introduce optimized GitHub Actions workflows for CI, release, and maintenance. Add development guides (DEVELOPMENT.md, CI_OPTIMIZATION.md), a cross-platform Makefile, and PowerShell dev script. Update project metadata, bump version to 0.1.3, and upgrade chrono dependency to 0.4.41. Major README improvements with detailed usage, contribution, and ecosystem info. Refactor and document local development workflow for efficiency.

+++ src/entity/scores/field/button.rs
@@ -27,11 +27,11 @@ impl StandardButtonSet {
         StandardButtonSet::from_bits(0)
     }
     pub fn is_down(&self, button: StandardButton) -> bool {
-        self.bits().bit(button.raw() as u32)
+        self.bits().bit(button.raw())
     }
     pub fn set_down(&self, button: StandardButton, is_down: bool) -> StandardButtonSet {
         let mut bits = self.bits();
-        bits.set_bit(button.raw() as u32, is_down);
+        bits.set_bit(button.raw(), is_down);
         StandardButtonSet::from_bits(bits)
     }
     pub fn press(&self, button: StandardButton) -> StandardButtonSet {
